修复编译错误：MemoryStream 容量参数需 int 而非 long
new MemoryStream(ftypBox.Length + mdatSize + newMoovSize) 传了 long，
而容量构造参数是 int。改为在求和处转 (int)，短视频体积远小于 2GB 无溢出。

+++ WeiboAlbumDownloader/Helpers/MotionPhotoHelper.cs
@@ -658,7 +658,8 @@ namespace WeiboAlbumDownloader.Helpers
             var ftypBox = new byte[ftypSize];
             Buffer.BlockCopy(data, 0, ftypBox, 0, (int)ftypSize); // ftyp 恒在偏移 0
 
-            using var ms = new MemoryStream(ftypBox.Length + mdatSize + newMoovSize);
+            // MemoryStream 容量参数为 int；短视频体积远小于 2GB，long 各长度转 int 无溢出
+            using var ms = new MemoryStream((int)(ftypBox.Length + mdatSize + newMoovSize));
             ms.Write(ftypBox, 0, ftypBox.Length);
             var mdatHdr = new byte[8];
             WriteUInt32BE(mdatHdr, 0, (uint)mdatSize);
